# docs/swagger.yaml
basePath: /api/v1
...
host: localhost:8080
info:
  contact:
    name: Librarium
    url: https://github.com/fireball1725/librarium
  description: Self-hosted library tracking API. All protected endpoints require a
    Bearer JWT in the Authorization header.
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
  termsOfService: http://localhost/
  title: Librarium API
  version: 26.0.0-dev
paths:
...
  /admin/users:
    get:
      description: Returns a paginated list of all instance users. Instance admin
        only.
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 20, max 100)
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              items:
                items:
                  type: object
                type: array
              page:
                type: integer
              per_page:
                type: integer
              total:
                type: integer
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "403":
          description: Forbidden
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: List all users (admin)
      tags:
      - admin
    post:
      consumes:
      - application/json
      description: Creates a new user account bypassing registration settings. Instance
        admin only.
      parameters:
      - description: Create user request
        in: body
        name: body
        required: true
        schema:
          properties:
            display_name:
              type: string
            email:
              type: string
            password:
              type: string
            username:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              created_at:
                type: string
              display_name:
                type: string
              email:
                type: string
              id:
                type: string
              is_active:
                type: boolean
              is_instance_admin:
                type: boolean
              last_login_at:
                type: string
              username:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "403":
          description: Forbidden
          schema:
            properties:
              error:
                type: string
            type: object
        "409":
          description: Conflict
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a user (admin)
      tags:
      - admin
  /admin/users/{id}:
    delete:
      description: Permanently deletes a user account. Instance admin only. Cannot
        delete yourself.
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "403":
          description: Forbidden
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a user (admin)
      tags:
      - admin
    patch:
      consumes:
      - application/json
      description: Partially updates a user's profile or admin/active status. Instance
        admin only.
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        schema:
          properties:
            display_name:
              type: string
            email:
              type: string
            is_active:
              type: boolean
            is_instance_admin:
              type: boolean
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              created_at:
                type: string
              display_name:
                type: string
              email:
                type: string
              id:
                type: string
              is_active:
                type: boolean
              is_instance_admin:
                type: boolean
              last_login_at:
                type: string
              username:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "403":
          description: Forbidden
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a user (admin)
      tags:
      - admin
...
securityDefinitions:
  BearerAuth:
    description: 'Enter: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"